Remove return value from upsert
There's no reason to clone the value here and changing the value type to
Arc<V> means that it really needs to be Arc<Mutex<V>> for any upsert
operation to work. It may be reasonable to return an indication of
whether or not the value was inserted or updated, but this works well
enough for now.